+++ docs/search/functions_2.js
@@ -1,8 +1,9 @@
 var searchData=
 [
   ['camera_0',['Camera',['../class_c_s_c_i441_1_1_camera.html#a2b44075a51d0d334a6a8c0742e42f040',1,'CSCI441::Camera']]],
-  ['computeshaderprogram_1',['ComputeShaderProgram',['../class_c_s_c_i441_1_1_compute_shader_program.html#a549bc6c5ca95c6817917ab1286503303',1,'CSCI441::ComputeShaderProgram::ComputeShaderProgram(const char *computeShaderFilename)'],['../class_c_s_c_i441_1_1_compute_shader_program.html#ab9147a8c3838cd5e76afd26967fba4cc',1,'CSCI441::ComputeShaderProgram::ComputeShaderProgram(const ComputeShaderProgram &amp;)=delete']]],
-  ['computeviewmatrix_2',['computeViewMatrix',['../class_c_s_c_i441_1_1_camera.html#a4480255c83e0f506e79fae3f565dd9f8',1,'CSCI441::Camera']]],
-  ['copytobuffer_3',['copyToBuffer',['../class_c_s_c_i441_1_1_uniform_buffer_object.html#ab640cb7aec972a4b4a0a7d4b88a36ab7',1,'CSCI441::UniformBufferObject']]],
-  ['copytooffset_4',['copyToOffset',['../class_c_s_c_i441_1_1_uniform_buffer_object.html#a93bfc9c16fb2d49f35dc360a9009437b',1,'CSCI441::UniformBufferObject']]]
+  ['checkopenglerrors_1',['checkOpenGLErrors',['../_open_g_l_utils_8hpp.html#a5e12bc3ac0ce7e01c50191b09c889a2c',1,'CSCI441::OpenGLUtils']]],
+  ['computeshaderprogram_2',['ComputeShaderProgram',['../class_c_s_c_i441_1_1_compute_shader_program.html#a549bc6c5ca95c6817917ab1286503303',1,'CSCI441::ComputeShaderProgram::ComputeShaderProgram(const char *computeShaderFilename)'],['../class_c_s_c_i441_1_1_compute_shader_program.html#ab9147a8c3838cd5e76afd26967fba4cc',1,'CSCI441::ComputeShaderProgram::ComputeShaderProgram(const ComputeShaderProgram &amp;)=delete']]],
+  ['computeviewmatrix_3',['computeViewMatrix',['../class_c_s_c_i441_1_1_camera.html#a4480255c83e0f506e79fae3f565dd9f8',1,'CSCI441::Camera']]],
+  ['copytobuffer_4',['copyToBuffer',['../class_c_s_c_i441_1_1_uniform_buffer_object.html#ab640cb7aec972a4b4a0a7d4b88a36ab7',1,'CSCI441::UniformBufferObject']]],
+  ['copytooffset_5',['copyToOffset',['../class_c_s_c_i441_1_1_uniform_buffer_object.html#a93bfc9c16fb2d49f35dc360a9009437b',1,'CSCI441::UniformBufferObject']]]
 ];