+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Revision History
 
+## v ?.?.? - ?? ??? ????
+- Fixed GPU memory leak of not deleting cube IBOs
+
 ## v 5.5.0 - 18 Oct 2024
 - TextureUtils::loadAndRegister2DTexture() can silence error message
 - MD5Model silences all but final texture fail messages during load