+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Revision History
 
-## v ?.?.? - ?? ??? ????
+## v 5.6.0 - 24 Oct 2024
 - Fixed GPU memory leak of not deleting cube IBOs
 - Add CSCI441::OpenGLUtils::checkOpenGLError() method to check the error queue
 - When requesting an OpenGL 4.3+ debug context, register debug callback